+++ configs/datamodule/mnist.yaml
