+++ configs/experiment/example_full.yaml
@@ -32,7 +32,7 @@ trainer:
   num_sanity_val_steps: 0
 
 model:
-  _target_: src.models.mnist_model.MNISTLitModel
+  _target_: src.models.mnist_module.MNISTLitModule
   input_size: 784
   lin1_size: 256
   lin2_size: 256
